requirement.txt
audeer==1.18.0
audiofile==1.1.0
bidict==0.22.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
eventlet==0.33.1
filelock==3.7.1
greenlet==1.1.2
huggingface-hub==0.8.1
idna==3.3
joblib==1.1.0
numpy==1.23.0
packaging==21.3
Pillow==9.1.1
PyAudio==0.2.11
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
python-engineio==4.3.2
python-socketio==5.6.0
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
sacremoses==0.0.53
sentencepiece==0.1.96
six==1.16.0
SoundFile==0.10.3.post1
sox==1.4.1
SpeechRecognition==3.8.1
tokenizers==0.12.1
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.2.0
urllib3==1.26.9
websocket-client==1.3.3
